# apps/api/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Abu Al-Nas API",
    "description": "Jordan e-commerce REST API — Arabic default, JOD currency.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/api/v1", "description": "API v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "My orders",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Order list" } }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Order detail",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Order" } }
      }
    },
    "/orders/{id}/invoice": {
      "get": {
        "tags": ["Orders"],
        "summary": "Invoice (HTML, Arabic RTL — print to PDF)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "text/html invoice" } }
      }
    },
    "/orders/{id}/invoice.pdf": {
      "get": {
        "tags": ["Orders"],
        "summary": "Invoice PDF download",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "application/pdf" } }
      }
    },
    "/admin/orders/{id}/invoice": {
      "get": {
        "tags": ["Admin"],
        "summary": "Admin invoice HTML",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "text/html" } }
      }
    },
...
  }
}